refactor protzilla.utilities
