--- protzilla/utilities/random.py
@@ -1,6 +1,0 @@
-from random import choices
-from string import ascii_letters
-
-
-def random_string():
-    return "".join(choices(ascii_letters, k=16))
