+++ tests/ui/test_views.py
@@ -9,7 +9,7 @@ sys.path.append(f"{UI_PATH}")
 from protzilla import data_preprocessing
 from protzilla.constants.paths import PROJECT_PATH, RUNS_PATH
 from protzilla.run import Run
-from protzilla.utilities.random import random_string
+from protzilla.utilities import random_string
 from ui.runs.views import active_runs, all_button_parameters
