+++ ui/runs/views.py
@@ -16,7 +16,7 @@ sys.path.append(f"{BASE_DIR}/..")
 
 from protzilla.run import Run
 from protzilla.run_helper import get_parameters
-from protzilla.utilities.memory import get_memory_usage
+from protzilla.utilities import get_memory_usage
 from ui.runs.fields import (
     make_current_fields,
     make_displayed_history,
